Remove Redirects navbar link & reorganize folders
