+++ pages/pages.json
@@ -68,10 +68,6 @@
             "path": "/admin/cms/pages"
           },
           {
-            "labelId": "pages.admin-menu-button.redirectList",
-            "path": "/admin/cms/redirects"
-          },
-          {
             "labelId": "pages.admin-menu-button.storefront",
             "path": "/admin/cms/storefront"
           },
