+++ react/locales/en.json
@@ -2,7 +2,6 @@
   "pages.admin-menu-button.displayConditions": "Display conditions",
   "pages.admin-menu-button.main": "Pages",
   "pages.admin-menu-button.pageList": "Page list",
-  "pages.admin-menu-button.redirectList": "Redirect list",
   "pages.admin-menu-button.storefront": "Storefront",
   "pages.conditions.device.any": "Any",
   "pages.conditions.device.desktop": "Desktop",
@@ -33,6 +32,7 @@
   "pages.admin.loading": "Loading",
   "pages.admin.pages.form.button.cancel": "Cancel",
   "pages.admin.pages.form.button.save": "Save",
+  "pages.admin.redirects.title": "Redirect list",
   "pages.admin.redirects.button.create": "New redirect",
   "pages.admin.redirects.emptyState": "No redirects created.",
   "pages.admin.redirects.form.button.back": "Back",
