+++ react/locales/es.json
@@ -2,7 +2,6 @@
   "pages.admin-menu-button.displayConditions": "Condiciones de exhibición",
   "pages.admin-menu-button.main": "Páginas",
   "pages.admin-menu-button.pageList": "Lista de páginas",
-  "pages.admin-menu-button.redirectList": "Lista de redirecciones",
   "pages.admin-menu-button.storefront": "Storefront",
   "pages.conditions.scope.route": "Esta ruta",
   "pages.conditions.scope.site": "Todo el sitio",
@@ -24,6 +23,7 @@
   "pages.admin.loading": "Cargando",
   "pages.admin.pages.form.button.cancel": "Cancelar",
   "pages.admin.pages.form.button.save": "Salvar",
+  "pages.admin.redirects.title": "Lista de redirecciones",
   "pages.admin.redirects.button.create": "Nueva redirección",
   "pages.admin.redirects.emptyState": "Ninguna redirección creada.",
   "pages.admin.redirects.form.button.back": "Volver",
