+++ react/locales/pt.json
@@ -2,7 +2,6 @@
   "pages.admin-menu-button.displayConditions": "Condições de exibição",
   "pages.admin-menu-button.main": "Páginas",
   "pages.admin-menu-button.pageList": "Lista de páginas",
-  "pages.admin-menu-button.redirectList": "Lista de redirecionamentos",
   "pages.admin-menu-button.storefront": "Storefront",
   "pages.conditions.scope.route": "Esta rota",
   "pages.conditions.scope.site": "Todo o site",
@@ -25,6 +24,7 @@
   "pages.admin.loading": "Carregando",
   "pages.admin.pages.form.button.cancel": "Cancelar",
   "pages.admin.pages.form.button.save": "Salvar",
+  "pages.admin.redirects.title": "Lista de redirecionamentos",
   "pages.admin.redirects.button.create": "Novo redirecionamento",
   "pages.admin.redirects.emptyState": "Nenhum redirecionamento criado.",
   "pages.admin.redirects.form.button.back": "Voltar",
